refactor: remove CustomFramebuffers class
