+++ src/main/java/ladysnake/satin/impl/CustomFormatFramebuffer.java
@@ -5,9 +5,12 @@ import com.mojang.blaze3d.platform.GlStateManager;
 import com.mojang.blaze3d.platform.TextureUtil;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.datafixers.types.templates.Check;
+import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gl.Framebuffer;
 import net.minecraft.client.gl.PostEffectProcessor;
 import net.minecraft.client.gl.SimpleFramebuffer;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL30;
 
 public class CustomFormatFramebuffer extends Framebuffer {
     private final int format;
@@ -17,6 +20,27 @@ public class CustomFormatFramebuffer extends Framebuffer {
         this.format = format;
         RenderSystem.assertOnRenderThreadOrInit();
         this.resize(width, height, getError);
+    }
+
+    public static Framebuffer createFramebuffer(int width, int height, String formatString) {
+        int format = decodeFormatString(formatString);
+        return new CustomFormatFramebuffer(format, width, height, true, MinecraftClient.IS_SYSTEM_MAC);
+    }
+
+    private static int decodeFormatString(String formatString) {
+        return switch (formatString) {
+            // unsigned normalised 8 bits
+            case "RGBA8" -> GL11.GL_RGBA8;
+            // unsigned normalised 16 bits
+            case "RGBA16" -> GL11.GL_RGBA16;
+            // float 16 bits
+            case "RGBA16F" -> GL30.GL_RGBA16F;
+            // float 32 bits
+            case "RGBA32F" -> GL30.GL_RGBA32F;
+            // we don't support un-normalised signed or unsigned integral formats here
+            // because it's not valid to blit between them and normalised/float formats
+            default -> throw new IllegalArgumentException("Unsupported texture format "+formatString);
+        };
     }
 
     /**
@@ -55,7 +79,8 @@ public class CustomFormatFramebuffer extends Framebuffer {
             GlStateManager._texParameter(GlConst.GL_TEXTURE_2D, GlConst.GL_TEXTURE_WRAP_S, GlConst.GL_CLAMP_TO_EDGE);
             GlStateManager._texParameter(GlConst.GL_TEXTURE_2D, GlConst.GL_TEXTURE_WRAP_T, GlConst.GL_CLAMP_TO_EDGE);
             GlStateManager._texImage2D(
-                    GlConst.GL_TEXTURE_2D, 0, format, this.textureWidth, this.textureHeight, 0, GlConst.GL_RGBA, GlConst.GL_UNSIGNED_BYTE, null
+                    // this ⬇️ is the only line changed from the super method
+                    GlConst.GL_TEXTURE_2D, 0, this.format, this.textureWidth, this.textureHeight, 0, GlConst.GL_RGBA, GlConst.GL_UNSIGNED_BYTE, null
             );
             GlStateManager._glBindFramebuffer(GlConst.GL_FRAMEBUFFER, this.fbo);
             GlStateManager._glFramebufferTexture2D(GlConst.GL_FRAMEBUFFER, GlConst.GL_COLOR_ATTACHMENT0, GlConst.GL_TEXTURE_2D, this.colorAttachment, 0);
